Added rest of code necessary for basic support page search
Added the methods count_key_words, sort_by_priority, and get_tweets_by_search. The last one is the method that the front end should call to get the tweets it needs for a given search. It takes in a profile name, a search, and a number of tweets requested and gives back a list of tweet-reply pairs. Each pair has the original tweet at index 0 and the reply at index 1. The tweets are the full tweet object for now, but could easily be changed to just the text if front-end requests this.